Update icons
Previously, to save time, this extension used the same icons as another Chrome extension I wrote. This commit updates the icon set to something more appropriate to the purpose of this extension.
The references to the icon files in manifest.json & scripts/service_worker.js have also, necessarily, been updated.